Move UML diagram test helpers to testSupport and fix test props
