+++ src/__tests__/components/canvas/UMLDiagram.save.test.tsx
@@ -1,20 +1,20 @@
-jest.mock('../../../utils/ecoreToUml', () => require('./diagramMockFactories').ecoreToUmlMock());
-jest.mock('../../../utils/saveMetaModelEcore', () => require('./diagramMockFactories').saveMetaModelEcoreMock());
-jest.mock('../../../utils/umlValidation', () => require('./diagramMockFactories').umlValidationMock());
-jest.mock('../../../utils/reactionFile', () => require('./diagramMockFactories').reactionFileMock());
-jest.mock('../../../components/canvas/UMLDiagramMinimap', () => require('./diagramMockFactories').umlDiagramMinimapMock());
-jest.mock('../../../components/flow/ReactionEditorModal', () => require('./diagramMockFactories').reactionEditorModalMock());
-jest.mock('../../../components/canvas/ReactionConfigPopup', () => require('./diagramMockFactories').reactionConfigPopupMock());
-jest.mock('../../../utils/umlDiagramGeometry', () => require('./diagramMockFactories').umlDiagramGeometryMock());
-jest.mock('../../../utils/umlClassLayout', () => require('./diagramMockFactories').umlClassLayoutMock());
-jest.mock('../../../utils/umlLayoutStorage', () => require('./diagramMockFactories').umlLayoutStorageMock());
+jest.mock('../../../utils/ecoreToUml', () => require('../../../testSupport/umlDiagram/mockFactories').ecoreToUmlMock());
+jest.mock('../../../utils/saveMetaModelEcore', () => require('../../../testSupport/umlDiagram/mockFactories').saveMetaModelEcoreMock());
+jest.mock('../../../utils/umlValidation', () => require('../../../testSupport/umlDiagram/mockFactories').umlValidationMock());
+jest.mock('../../../utils/reactionFile', () => require('../../../testSupport/umlDiagram/mockFactories').reactionFileMock());
+jest.mock('../../../components/canvas/UMLDiagramMinimap', () => require('../../../testSupport/umlDiagram/mockFactories').umlDiagramMinimapMock());
+jest.mock('../../../components/flow/ReactionEditorModal', () => require('../../../testSupport/umlDiagram/mockFactories').reactionEditorModalMock());
+jest.mock('../../../components/canvas/ReactionConfigPopup', () => require('../../../testSupport/umlDiagram/mockFactories').reactionConfigPopupMock());
+jest.mock('../../../utils/umlDiagramGeometry', () => require('../../../testSupport/umlDiagram/mockFactories').umlDiagramGeometryMock());
+jest.mock('../../../utils/umlClassLayout', () => require('../../../testSupport/umlDiagram/mockFactories').umlClassLayoutMock());
+jest.mock('../../../utils/umlLayoutStorage', () => require('../../../testSupport/umlDiagram/mockFactories').umlLayoutStorageMock());
 
 import React, { createRef } from 'react';
 import { act, screen, waitFor } from '@testing-library/react';
 import { saveMetaModelEcore } from '../../../utils/saveMetaModelEcore';
 import type { UMLDiagramHandle } from '../../../components/canvas/UMLDiagram';
-import { SIMPLE_ECORE } from './diagramTestFixtures';
-import { renderDiagram } from './diagramTestUtils';
+import { SIMPLE_ECORE } from '../../../testSupport/umlDiagram/fixtures';
+import { renderDiagram } from '../../../testSupport/umlDiagram/renderUtils';
 
 describe('UMLDiagram save', () => {
   it('workspace save calls onSaved without hitting the library API', async () => {
@@ -44,3 +44,4 @@ describe('UMLDiagram save', () => {
     expect(screen.getByText('Saved to project')).toBeInTheDocument();
   });
 });
+
